Extracted components from CalendarScreen
